refactor: update design menu layout
