+++ package.json
@@ -27,6 +27,7 @@
     "@vueuse/core": "^10.2.1",
     "@vueuse/nuxt": "^10.2.1",
     "autoprefixer": "^10.4.14",
+    "checkly": "latest",
     "happy-dom": "^9.20.3",
     "nuxt": "^3.6.5",
     "nuxt-headlessui": "^1.1.4",
@@ -35,8 +36,7 @@
     "postcss": "^8.4.27",
     "ts-node": "latest",
     "typescript": "latest",
-    "vitest": "^0.33.0",
-    "checkly": "latest"
+    "vitest": "^0.33.0"
   },
   "dependencies": {
     "@firebase/app-types": "^0.9.0",
